fixup! Extract workflow and improve argument parsing.
